+++ testsuite/test_valuemapping.py
@@ -145,7 +145,7 @@ class Test_ValueMapping(object):
     @staticmethod
     def assertOutsideValueMap(vm, value):
         """
-        Test vm.tovalues Exception
+        Test vm.tovalues() Exception
         """
         with pytest.raises(ValueError) as exc_info:
             vm.tovalues(value)
@@ -153,6 +153,17 @@ class Test_ValueMapping(object):
         exc_msg = str(exc)
         assert re.match("Element value outside of the set defined by "
                         "ValueMap.*", exc_msg) is not None
+
+    def assertOutsideValues(self, vm, value):
+        """
+        Test vm.tobinary() Exception
+        """
+        with pytest.raises(ValueError) as exc_info:
+            vm.tobinary(value)
+        exc = exc_info.value
+        exc_msg = str(exc)
+        assert re.match("Values string outside of the set defined by "
+                        "Values.*", exc_msg) is not None
 
     def test_empty(self, element_kind, server_arg, integer_type):
         # pylint: disable=redefined-outer-name
@@ -164,6 +175,7 @@ class Test_ValueMapping(object):
                                     valuemap, values)
 
         self.assertOutsideValueMap(vm, 0)
+        self.assertOutsideValues(vm, 'x')
 
     def test_attrs_property(self, server_arg, integer_type):
         # pylint: disable=redefined-outer-name
@@ -310,6 +322,22 @@ class Test_ValueMapping(object):
         assert isinstance(exc, TypeError)
         assert exc.args[0].startswith('Element value is not an integer type')
 
+    def test_invalid_tobinary_type(self, element_kind, server_arg,
+                                   integer_type):
+        # pylint: disable=redefined-outer-name
+        """Test tobinary() with invalid type"""
+        valuemap = ['0']
+        values = ['zero']
+
+        vm = self.setup_for_element(element_kind, server_arg, integer_type,
+                                    valuemap, values)
+
+        with pytest.raises(Exception) as exc_info:
+            vm.tobinary(0)
+        exc = exc_info.value
+        assert isinstance(exc, TypeError)
+        assert exc.args[0].startswith('Values string is not a string type')
+
     def test_no_values_qualifier(self, element_kind, server_arg, integer_type):
         # pylint: disable=redefined-outer-name
         """Test element without Values qualifier"""
@@ -327,6 +355,7 @@ class Test_ValueMapping(object):
         # pylint: disable=redefined-outer-name
         """Test default if no ValueMap qualifier is defined"""
         valuemap = None
+        exp_valuemap = [0, 1, 2]
         values = ['zero', 'one', 'two']
 
         vm = self.setup_for_element(element_kind, server_arg, integer_type,
@@ -338,10 +367,22 @@ class Test_ValueMapping(object):
         assert vm.tovalues(2) == 'two'
         self.assertOutsideValueMap(vm, 3)
 
+        self.assertOutsideValues(vm, 'x')
+        assert vm.tobinary('zero') == 0
+        assert vm.tobinary('one') == 1
+        assert vm.tobinary('two') == 2
+
+        exp_items = list(zip(exp_valuemap, values))
+        items = []
+        for item in vm.items():
+            items.append(item)
+        assert items == exp_items
+
     def test_zero(self, element_kind, server_arg, integer_type):
         # pylint: disable=redefined-outer-name
         """Test valuemap with value 0"""
         valuemap = ['0']
+        exp_valuemap = [0]
         values = ['zero']
 
         vm = self.setup_for_element(element_kind, server_arg, integer_type,
@@ -350,10 +391,20 @@ class Test_ValueMapping(object):
         assert vm.tovalues(0) == 'zero'
         self.assertOutsideValueMap(vm, 1)
 
+        self.assertOutsideValues(vm, 'x')
+        assert vm.tobinary('zero') == 0
+
+        exp_items = list(zip(exp_valuemap, values))
+        items = []
+        for item in vm.items():
+            items.append(item)
+        assert items == exp_items
+
     def test_one(self, element_kind, server_arg, integer_type):
         # pylint: disable=redefined-outer-name
         """Test valuemap with value 1"""
         valuemap = ['1']
+        exp_valuemap = [1]
         values = ['one']
 
         vm = self.setup_for_element(element_kind, server_arg, integer_type,
@@ -363,9 +414,19 @@ class Test_ValueMapping(object):
         assert vm.tovalues(1) == 'one'
         self.assertOutsideValueMap(vm, 2)
 
-    def test_one_cimtype(self, element_kind, server_arg, integer_type):
+        self.assertOutsideValues(vm, 'x')
+        assert vm.tobinary('one') == 1
+
+        exp_items = list(zip(exp_valuemap, values))
+        items = []
+        for item in vm.items():
+            items.append(item)
+        assert items == exp_items
+
+    def test_tovalues_with_cimtype(self, element_kind, server_arg,
+                                   integer_type):
         # pylint: disable=redefined-outer-name
-        """Test valuemap with value cimtype(1)"""
+        """Test tovalues() with argument of CIM type"""
         valuemap = ['1']
         values = ['one']
 
@@ -375,10 +436,23 @@ class Test_ValueMapping(object):
         cimtype = type_from_name(integer_type)
         assert vm.tovalues(cimtype(1)) == 'one'
 
+    def test_tobinary_with_unicode(self, element_kind, server_arg,
+                                   integer_type):
+        # pylint: disable=redefined-outer-name
+        """Test tobinary() with argument unicode string"""
+        valuemap = ['1']
+        values = ['one']
+
+        vm = self.setup_for_element(element_kind, server_arg, integer_type,
+                                    valuemap, values)
+
+        assert vm.tobinary(u'one') == 1
+
     def test_singles(self, element_kind, server_arg, integer_type):
         # pylint: disable=redefined-outer-name
         """Test valuemap with multiple non-range values"""
         valuemap = ['0', '1', '9']
+        exp_valuemap = [0, 1, 9]
         values = ['zero', 'one', 'nine']
 
         vm = self.setup_for_element(element_kind, server_arg, integer_type,
@@ -391,10 +465,22 @@ class Test_ValueMapping(object):
         assert vm.tovalues(9) == 'nine'
         self.assertOutsideValueMap(vm, 10)
 
+        self.assertOutsideValues(vm, 'x')
+        assert vm.tobinary('zero') == 0
+        assert vm.tobinary('one') == 1
+        assert vm.tobinary('nine') == 9
+
+        exp_items = list(zip(exp_valuemap, values))
+        items = []
+        for item in vm.items():
+            items.append(item)
+        assert items == exp_items
+
     def test_singles_ranges(self, element_kind, server_arg, integer_type):
         # pylint: disable=redefined-outer-name
         """Test valuemap with ranges"""
         valuemap = ['0', '1', '2..4', '..6', '7..', '9']
+        exp_valuemap = [0, 1, (2, 4), (5, 6), (7, 8), 9]
         values = ['zero', 'one', 'two-four', 'five-six', 'seven-eight', 'nine']
 
         vm = self.setup_for_element(element_kind, server_arg, integer_type,
@@ -412,10 +498,25 @@ class Test_ValueMapping(object):
         assert vm.tovalues(9) == 'nine'
         self.assertOutsideValueMap(vm, 10)
 
+        self.assertOutsideValues(vm, 'x')
+        assert vm.tobinary('zero') == 0
+        assert vm.tobinary('one') == 1
+        assert vm.tobinary('two-four') == (2, 4)
+        assert vm.tobinary('five-six') == (5, 6)
+        assert vm.tobinary('seven-eight') == (7, 8)
+        assert vm.tobinary('nine') == 9
+
+        exp_items = list(zip(exp_valuemap, values))
+        items = []
+        for item in vm.items():
+            items.append(item)
+        assert items == exp_items
+
     def test_unclaimed(self, element_kind, server_arg, integer_type):
         # pylint: disable=redefined-outer-name
         """Test valuemap with unclaimed marker '..'"""
         valuemap = ['..']
+        exp_valuemap = [None]
         values = ['unclaimed']
 
         vm = self.setup_for_element(element_kind, server_arg, integer_type,
@@ -425,11 +526,20 @@ class Test_ValueMapping(object):
         assert vm.tovalues(1) == 'unclaimed'
         assert vm.tovalues(2) == 'unclaimed'
 
+        assert vm.tobinary('unclaimed') is None
+
+        exp_items = list(zip(exp_valuemap, values))
+        items = []
+        for item in vm.items():
+            items.append(item)
+        assert items == exp_items
+
     def test_singles_ranges_unclaimed(self, element_kind, server_arg,
                                       integer_type):
         # pylint: disable=redefined-outer-name
         """Test valuemap with combination of singles, ranges, unclaimed"""
         valuemap = ['0', '1', '2..4', '..6', '7..', '9', '..']
+        exp_valuemap = [0, 1, (2, 4), (5, 6), (7, 8), 9, None]
         values = ['zero', 'one', 'two-four', 'five-six', 'seven-eight', 'nine',
                   'unclaimed']
 
@@ -449,11 +559,26 @@ class Test_ValueMapping(object):
         assert vm.tovalues(10) == 'unclaimed'
         assert vm.tovalues(11) == 'unclaimed'
 
+        assert vm.tobinary('zero') == 0
+        assert vm.tobinary('one') == 1
+        assert vm.tobinary('two-four') == (2, 4)
+        assert vm.tobinary('five-six') == (5, 6)
+        assert vm.tobinary('seven-eight') == (7, 8)
+        assert vm.tobinary('nine') == 9
+        assert vm.tobinary('unclaimed') is None
+
+        exp_items = list(zip(exp_valuemap, values))
+        items = []
+        for item in vm.items():
+            items.append(item)
+        assert items == exp_items
+
     def test_singles_ranges_unclaimed2(self, element_kind, server_arg,
                                        integer_type):
         # pylint: disable=redefined-outer-name
         """Test singles, ranges, unclaimed"""
         valuemap = ['0', '2..4', '..6', '7..', '9', '..']
+        exp_valuemap = [0, (2, 4), (5, 6), (7, 8), 9, None]
         values = ['zero', 'two-four', 'five-six', 'seven-eight', 'nine',
                   'unclaimed']
 
@@ -473,6 +598,19 @@ class Test_ValueMapping(object):
         assert vm.tovalues(10) == 'unclaimed'
         assert vm.tovalues(11) == 'unclaimed'
 
+        assert vm.tobinary('zero') == 0
+        assert vm.tobinary('two-four') == (2, 4)
+        assert vm.tobinary('five-six') == (5, 6)
+        assert vm.tobinary('seven-eight') == (7, 8)
+        assert vm.tobinary('nine') == 9
+        assert vm.tobinary('unclaimed') is None
+
+        exp_items = list(zip(exp_valuemap, values))
+        items = []
+        for item in vm.items():
+            items.append(item)
+        assert items == exp_items
+
     def test_min_max_single(self, element_kind, server_arg, integer_type):
         # pylint: disable=redefined-outer-name
         """Test valuemap with single min and max values of integer type"""
@@ -483,6 +621,8 @@ class Test_ValueMapping(object):
         valuemap = []
         valuemap.append(str(minvalue))
         valuemap.append(str(maxvalue))
+
+        exp_valuemap = [minvalue, maxvalue]
 
         values = []
         for v in valuemap:
@@ -499,9 +639,20 @@ class Test_ValueMapping(object):
         assert vm.tovalues(maxvalue) == values[1]
         self.assertOutsideValueMap(vm, maxvalue + 1)
 
-    def test_min_max_closed_range(self, element_kind, server_arg, integer_type):
+        self.assertOutsideValues(vm, 'x')
+        assert vm.tobinary(values[0]) == minvalue
+        assert vm.tobinary(values[1]) == maxvalue
+
+        exp_items = list(zip(exp_valuemap, values))
+        items = []
+        for item in vm.items():
+            items.append(item)
+        assert items == exp_items
+
+    def test_min_max_closed_ranges(self, element_kind, server_arg,
+                                   integer_type):
         # pylint: disable=redefined-outer-name
-        """Test valuemap with closed range of min and max of integer type"""
+        """Test valuemap with two closed ranges at min and max of int type"""
 
         minvalue = type_from_name(integer_type).minvalue
         maxvalue = type_from_name(integer_type).maxvalue
@@ -509,6 +660,8 @@ class Test_ValueMapping(object):
         valuemap = []
         valuemap.append('%s..%s' % (minvalue, minvalue + 2))
         valuemap.append('%s..%s' % (maxvalue - 2, maxvalue))
+
+        exp_valuemap = [(minvalue, minvalue + 2), (maxvalue - 2, maxvalue)]
 
         values = []
         for v in valuemap:
@@ -529,6 +682,16 @@ class Test_ValueMapping(object):
         assert vm.tovalues(maxvalue) == values[1]
         self.assertOutsideValueMap(vm, maxvalue + 1)
 
+        self.assertOutsideValues(vm, 'x')
+        assert vm.tobinary(values[0]) == (minvalue, minvalue + 2)
+        assert vm.tobinary(values[1]) == (maxvalue - 2, maxvalue)
+
+        exp_items = list(zip(exp_valuemap, values))
+        items = []
+        for item in vm.items():
+            items.append(item)
+        assert items == exp_items
+
     def test_min_max_open_range1(self, element_kind, server_arg, integer_type):
         # pylint: disable=redefined-outer-name
         """Test valuemap with open range 1 of min and max of integer type"""
@@ -541,6 +704,9 @@ class Test_ValueMapping(object):
         valuemap.append('..%s' % (minvalue + 2))
         valuemap.append('%s..' % (maxvalue - 2))
         valuemap.append('%s' % maxvalue)
+
+        exp_valuemap = [minvalue, (minvalue + 1, minvalue + 2),
+                        (maxvalue - 2, maxvalue - 1), maxvalue]
 
         values = []
         for v in valuemap:
@@ -561,6 +727,18 @@ class Test_ValueMapping(object):
         assert vm.tovalues(maxvalue) == values[3]
         self.assertOutsideValueMap(vm, maxvalue + 1)
 
+        self.assertOutsideValues(vm, 'x')
+        assert vm.tobinary(values[0]) == minvalue
+        assert vm.tobinary(values[1]) == (minvalue + 1, minvalue + 2)
+        assert vm.tobinary(values[2]) == (maxvalue - 2, maxvalue - 1)
+        assert vm.tobinary(values[3]) == maxvalue
+
+        exp_items = list(zip(exp_valuemap, values))
+        items = []
+        for item in vm.items():
+            items.append(item)
+        assert items == exp_items
+
     def test_min_max_open_range2(self, element_kind, server_arg, integer_type):
         # pylint: disable=redefined-outer-name
         """Test valuemap with open range 2 of min and max of integer type"""
@@ -571,6 +749,9 @@ class Test_ValueMapping(object):
         valuemap = []
         valuemap.append('..%s' % (minvalue + 2))
         valuemap.append('%s..' % (maxvalue - 2))
+
+        exp_valuemap = [(minvalue, minvalue + 2),
+                        (maxvalue - 2, maxvalue)]
 
         values = []
         for v in valuemap:
@@ -590,3 +771,13 @@ class Test_ValueMapping(object):
         assert vm.tovalues(maxvalue - 1) == values[1]
         assert vm.tovalues(maxvalue) == values[1]
         self.assertOutsideValueMap(vm, maxvalue + 1)
+
+        self.assertOutsideValues(vm, 'x')
+        assert vm.tobinary(values[0]) == (minvalue, minvalue + 2)
+        assert vm.tobinary(values[1]) == (maxvalue - 2, maxvalue)
+
+        exp_items = list(zip(exp_valuemap, values))
+        items = []
+        for item in vm.items():
+            items.append(item)
+        assert items == exp_items